XenServer Dashboards first commit
-Added the following Dashboards for Xenserver:
- xs-host-info.json
Host Information Dashboard that includes the following information:
- Hostname
- Host UUID
- Host Version
- Host CPU
- Host Memory
- Host Storage
Also Includes a graph for the number of VMs and PIFs.

- xs-pool-info-overall.json
Overall Pool Information Dashboard that includes the following information:
Information about the Hosts running on each pool, with the option to select a specific host to view more detailed information.
Includes a graph for the number of Hosts running on the pool.

- xs-pool-info.json
Pool Information Dashboard that includes the following information:
Information about the Hosts running on each pool, with the option to select a specific host to view more detailed information.
Includes a graph for the number of Hosts running on the pool.

- xs-vm-info.json
VM Information Dashboard that includes the following information:
Information about the VMs running on the Host, with the option to select a specific VM that includes information about the VM's CPU, Memory, IOPS, and Network.